assignment/assignment2/requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
pyspark
xgboost
psycopg2-binary